Node_Hardware/pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "node-hardware-mcp"
version = "0.1.0"
description = "Node Hardware MCP server implementation using Model Context Protocol"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "mcp[cli]>=0.1.0",
  "pytest-asyncio>=1.0.0",
  "python-dotenv>=1.0.0",
  "psutil>=5.9.0",
  "fastapi>=0.95.0",
  "uvicorn>=0.21.0",
  "pydantic>=1.10.0",
  "pytest>=7.2.0",
  "requests>=2.28.0"
]

[project.optional-dependencies]
test = ["pytest>=7.0.0", "httpx"]

[project.scripts]
node-hardware-mcp = "node_hardware.server:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["node_hardware*"]

[tool.setuptools.package-dir]
"node_hardware" = "src/node_hardware"

[dependency-groups]
dev = [
    "pytest>=8.4.0",
]

[tool.uv.workspace]
members = [
    "node-hardware-mcp",
]